project/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
firebase-admin
cloudinary
SQLAlchemy
psycopg2-binary
fastapi-cors
python-multipart
fastapi
uvicorn
pydantic
langchain
sentence-transformers
faiss-cpu